Bubble Chart improvements

+++ Charts/Classes/Data/BubbleChartDataEntry.swift
@@ -14,12 +14,12 @@ import Foundation
 public class BubbleChartDataEntry: ChartDataEntry
 {
     /// The size of the bubble.
-    public var size = Float(0.0)
+    public var size = CGFloat(0.0)
     
     /// :xIndex: The index on the x-axis.
     /// :val: The value on the y-axis.
     /// :size: The size of the bubble.
-    public init(xIndex: Int, value: Float, size: Float)
+    public init(xIndex: Int, value: Float, size: CGFloat)
     {
         super.init(value: value, xIndex: xIndex)
         
@@ -30,7 +30,7 @@ public class BubbleChartDataEntry: ChartDataEntry
     /// :val: The value on the y-axis.
     /// :size: The size of the bubble.
     /// :data: Spot for additional data this Entry represents.
-    public init(xIndex: Int, value: Float, size: Float, data: AnyObject?)
+    public init(xIndex: Int, value: Float, size: CGFloat, data: AnyObject?)
     {
         super.init(value: value, xIndex: xIndex, data: data)
